Fix date formatter and improve date usage

+++ app/src/main/java/com/example/dotametrics/data/model/players/heroes/PlayerHeroResult.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class PlayerHeroResult(
     @SerializedName("hero_id") var heroId: String? = null,
-    @SerializedName("last_played") var lastPlayed: Int? = null,
+    @SerializedName("last_played") var lastPlayed: Long? = null,
     @SerializedName("games") var games: Int? = null,
     @SerializedName("win") var win: Int? = null,
     @SerializedName("with_games") var withGames: Int? = null,
